+++ test/auth-test.js
@@ -9,6 +9,14 @@ cds.test(__dirname + "/_env");
 
 let request;
 
+const validAuth = `Basic ${Buffer.from(
+  `${cds.requires.auth.users.alice.id}:${cds.requires.auth.users.alice.password}`
+).toString("base64")}`;
+
+const invalidAuth = `Basic ${Buffer.from(
+  `${cds.requires.auth.users.bob.id}:${cds.requires.auth.users.bob.password}`
+).toString("base64")}`;
+
 describe("auth", () => {
   beforeAll(async () => {
     await global._init;
@@ -22,38 +30,31 @@ describe("auth", () => {
     expect(response.status).toEqual(401);
     expect(response.headers["www-authenticate"]).toEqual('Basic realm="Users"');
 
-    let authorization = `Basic ${Buffer.from(
-      `${cds.requires.auth.users.bob.id}:${cds.requires.auth.users.bob.password}`
-    ).toString("base64")}`;
     response = await util.callRead(request, "/v2/auth/$metadata", {
       accept: "application/xml",
-      Authorization: authorization,
+      Authorization: invalidAuth,
     });
     expect(response.status).toEqual(403);
 
-    authorization = `Basic ${Buffer.from(
-      `${cds.requires.auth.users.alice.id}:${cds.requires.auth.users.alice.password}`
-    ).toString("base64")}`;
     response = await util.callRead(request, "/v2/auth/$metadata", {
       accept: "application/xml",
-      Authorization: authorization,
+      Authorization: validAuth,
     });
     expect(response.status).toEqual(200);
 
-    authorization = `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c`;
     response = await util.callRead(request, "/v2/auth/$metadata", {
       accept: "application/xml",
-      Authorization: authorization,
+      Authorization:
+        "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
     });
     expect(response.status).toEqual(401);
   });
 
   it("GET $metadata invalid auth", async () => {
     const consoleSpy = jest.spyOn(console, "error");
-    let authorization = `Bearer xyz`;
     let response = await util.callRead(request, "/v2/auth/$metadata", {
       accept: "application/xml",
-      Authorization: authorization,
+      Authorization: "Bearer xyz",
     });
     expect(response.status).toEqual(401);
     expect(consoleSpy).toHaveBeenCalledWith(
@@ -61,15 +62,36 @@ describe("auth", () => {
       "Authorization:",
       expect.objectContaining(new Error("Invalid JWT token"))
     );
+    response = await util.callRead(request, "/v2/auth/$metadata", {
+      accept: "application/xml",
+      Authorization: validAuth,
+    });
+    expect(response.status).toEqual(200);
+  });
+
+  it("GET service root invalid auth", async () => {
+    const consoleSpy = jest.spyOn(console, "error");
+    let response = await util.callRead(request, "/v2/auth/", {
+      accept: "application/xml",
+      Authorization: "Bearer xyz",
+    });
+    expect(response.status).toEqual(401);
+    expect(consoleSpy).toHaveBeenCalledWith(
+      "[cov2ap] -",
+      "Authorization:",
+      expect.objectContaining(new Error("Invalid JWT token"))
+    );
+    response = await util.callRead(request, "/v2/auth/", {
+      accept: "application/xml",
+      Authorization: validAuth,
+    });
+    expect(response.status).toEqual(200);
   });
 
   it("GET $metadata check response correlation", async () => {
-    const authorization = `Basic ${Buffer.from(
-      `${cds.requires.auth.users.alice.id}:${cds.requires.auth.users.alice.password}`
-    ).toString("base64")}`;
     const response = await util.callRead(request, "/v2/auth/$metadata", {
       accept: "application/xml",
-      Authorization: authorization,
+      Authorization: validAuth,
     });
     expect(response.status).toEqual(200);
     expect(response.headers["x-request-id"]).toBeDefined();