+++ test/main-test.js
@@ -2826,7 +2826,7 @@ describe("main", () => {
     });
   });
 
-  it("GET with x-forwarded-path header", async () => {
+  it("GET with x-forwarded headers", async () => {
     let response = await util.callWrite(request, "/v2/main/Header", {
       name: "Test",
     });