+++ src/index.js
@@ -319,17 +319,15 @@ function cov2ap(options = {}) {
   router.get(`/${path}/*\\$metadata`, async (req, res) => {
     let serviceValid = true;
     try {
-      let metadataUrlPath = targetUrl(req);
-      if (metadataUrlPath.endsWith("/$metadata")) {
-        metadataUrlPath = metadataUrlPath.substring(0, metadataUrlPath.length - 9);
-      }
-      const metadataTargetUrl = target + metadataUrlPath;
+      const urlPath = targetUrl(req);
+      const metadataUrl = URL.parse(urlPath, true);
+      const serviceUrl = target + metadataUrl.pathname.substring(0, metadataUrl.pathname.length - 9);
       // Trace
       traceRequest(req, "Request", req.method, req.originalUrl, req.headers, req.body);
-      traceRequest(req, "ProxyRequest", req.method, metadataUrlPath, req.headers, req.body);
+      traceRequest(req, "ProxyRequest", req.method, urlPath, req.headers, req.body);
 
       const result = await Promise.all([
-        fetch(metadataTargetUrl, {
+        fetch(serviceUrl, {
           method: "GET",
           headers: {
             ...propagateHeaders(req),