+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Support new integer types
 - Fallback target `auto` to `default` target until dynamic target/port assignment is available
+- Prevent loading OData V4 $metadata when OData V2 $metadata is requested
 
 ## Version 1.9.10 - 2022-10-04