+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Support new integer types
+- Support new CDS integer types
 - Fallback target `auto` to `default` target until dynamic target/port assignment is available
 - Prevent loading OData V4 $metadata when OData V2 $metadata is requested